refactor: clear the residual new-code SonarCloud findings from the first pass
The first remediation pass overcounted returns on _corroborated_subject (4,
not 3) and pushed _gate_checks.py's line count over budget by adding a helper
inline. Extract _service_materialization_binding to bring the return count
under the threshold, split the offline _NoStartBackend conformance stub into
_gate_no_start_backend.py (re-exported for its existing external import
sites), and mark _build_seed_script static (it never touches self).
